refactor: use latest change from directory for generating image tag
The actual behaviour of the rock repositories is to check for changes in the directory where
a rockcraft.yaml file lives; for any changes detected, the container image will be published.
The get-image-names-from-rockcraft.yaml should also follow that pattern for correctly
generating the tags of the container images that will eventually be scanned.