ci: create a new workflow for getting charm paths and building them with cache (#88)
* ci: create a new workflow for getting charm paths and building them with cache

This new workflow uses the  as the central piece for building charms in dedicated machines and uploading them as artefacts. This is beneficial because it will allow the CI of pull requests to use charm artefacts in different operations,
such as testing and publishing, this change also helps reducing the build time as the
workflow is set to use charmcarft cache.

Fixes #87