chore: keep only setting for disabling renovate (#118)
Follow up of #117

In the end the above PR didn't disable the renovate PRs in workflows that are extending pointing to this central schema. We'll try to keep the only settings in defaults.json to be the ones that disable renovate.

If this also doesn't work then we'll need to disable renovate in every dedicated repo instead.